[FrameworkBundle] Adding the redirect method back to the base controller
Some question whether or not the base Controller should be included at all. I think it absolutely must be included because it's important for beginners and for rapid development of smaller features/applications (and rapid development is good for beginners).

So, assuming that we *do* like the base Controller, we should really use it to its fullest potential - making the lives of developers as easy as possible.