Fixed the danish to be the same in all the validations..
English:
The value is not a country.

Danish translation in the latest commit:
Denne værdi er ikke et land.

Danish translation in the first commit and mine version:
Værdien er ikke et land

So this commit is simply to make the danish translation the same, and
not two different expressions..