removed unneeded check
It is already checked in the `isEnabled()` method of the parent
`ServerCommand` class if the current PHP build is HHVM and the
`server:start` command is never enabled then. Thus, it's not needed
to check for HHVM on every command execution.

This was pointed out by @stof in #12489 for the `server:run` command.